fix extra

+++ src/main/java/com/keyin/hello/GreetingController.java
@@ -1,4 +1,4 @@
-//GreetingController Mohamed
+//GreetingController
 package com.keyin.hello;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +23,7 @@ public class GreetingController {
     }
 
     @GetMapping("greeting/{index}")
-    public Greeting getGreeting(@PathVariable Integer index) {
+    public Greeting getGreeting(@PathVariable Long index) {  // Changed Integer to Long to match the ID type in the entity
         return greetingService.getGreeting(index);
     }
 
@@ -33,12 +33,12 @@ public class GreetingController {
     }
 
     @PutMapping("greeting/{index}")
-    public Greeting updateGreeting(@PathVariable Integer index, @RequestBody Greeting updatedGreeting) {
+    public Greeting updateGreeting(@PathVariable Long index, @RequestBody Greeting updatedGreeting) {  // Changed Integer to Long to match the ID type in the entity
         return greetingService.updateGreeting(index, updatedGreeting);
     }
 
     @DeleteMapping("greeting/{index}")
-    public void deleteGreeting(@PathVariable Integer index) {
+    public void deleteGreeting(@PathVariable Long index) {  // Changed Integer to Long to match the ID type in the entity
         greetingService.deleteGreeting(index);
     }
 }
